+++ lib/rb_sdl2/surface/pixel_format.rb
@@ -12,30 +12,27 @@ module RbSDL2
       end
 
       class << self
-        require_relative '../pixel_format_enum'
-
-        def new(format)
-          ptr = PixelFormatPointer.new(::SDL::AllocFormat(PixelFormatEnum.to_num(format)))
-          raise RbSDL2Error if ptr.null?
-          super(ptr)
-        end
-
-        def to_ptr(ptr)
-          obj = allocate
-          obj.__send__(:initialize, PixelFormatPointer.to_ptr(ptr))
-          obj
-        end
+        # アプリケーションが PixelFormat を作成する必要がないためコンストラクターを実装しない。
+        # FFI::AutoPointer によるポインター管理を行っているのは Surface 側で PixelFormat を公開しているため。
       end
 
       def initialize(ptr)
-        @st = ::SDL::PixelFormat.new(ptr)
+        @st = ::SDL::PixelFormat.new(PixelFormatPointer.to_ptr(ptr))
       end
 
       def ==(other)
-        other.respond_to?(:to_ptr) && to_ptr == other.to_ptr
+        other.respond_to?(:to_ptr) && other.to_ptr == to_ptr
       end
 
-      def alpha_mask? = @st[:Amask] > 0
+      def a_mask = @st[:Amask]
+
+      def b_mask = @st[:Bmask]
+
+      def g_mask = @st[:Gmask]
+
+      def r_mask = @st[:Rmask]
+
+      def a_mask? = @st[:Amask] > 0
 
       def bits_per_pixel = @st[:BitsPerPixel]
       alias bpp bits_per_pixel
@@ -50,7 +47,7 @@ module RbSDL2
       # indexed format のときはパレット番号を戻す。
       def pack_color(color)
         r, g, b, a = color
-        if alpha_mask?
+        if a_mask?
           ::SDL.MapRGBA(self, r, g, b, a || ::SDL::ALPHA_OPAQUE)
         else
           ::SDL.MapRGB(self, r, g, b)
@@ -59,28 +56,32 @@ module RbSDL2
 
       require_relative '../palette'
 
+      # パレットがある場合は Palette インスタンスを戻します。ない場合は nil を戻します。
       def palette
-        # パレットは参照カウンターで生存の保証がある。
-        # Ruby 側がパレットを保持している限り同一アドレスに違うパレットが作成されることはない。
-        # SDL では PixelFormat の palette メンバーは（行儀よく SetPixelFormatPalette を使う場合は）
-        # 後から NULL に書き換わることはない。
-        (ptr = @st[:palette]) == @palette&.to_ptr ? @palette : @palette = Palette.to_ptr(ptr)
+        Palette.to_ptr(@st[:palette]) unless @st[:palette].null?
       end
 
       def palette=(pal)
-        err = ::SDL.SetPixelFormatPalette(self, pal)
-        raise RbSDL2Error if err < 0
-        @palette = nil
+        # SDL_SetPixelFormatPalette() はパレット・ポインターが NULL かどうかチェックしていない。
+        if Palette === pal && !pal.to_ptr.null?
+          err = ::SDL.SetPixelFormatPalette(self, pal)
+          raise RbSDL2Error if err < 0
+        else
+          raise ArgumentError, "pointer is NULL"
+        end
       end
+
+      def palette? = !@st[:palette].null?
 
       def to_ptr = @st.to_ptr
 
       # indexed format のときはパレット番号を引数へ与える。
-      def unpack_color(pixel)
-        if alpha_mask?
-          ::SDL.GetRGBA(pixel, self, *Array.new(4) { ::FFI::MemoryPointer.new(:uint8) })
+      def unpack_pixel(num)
+        color = Array.new(a_mask? ? 4 : 3) { ::FFI::MemoryPointer.new(:uint8) }
+        if a_mask?
+          ::SDL.GetRGBA(num, self, *color)
         else
-          ::SDL.GetRGB(pixel, self, *Array.new(3) { ::FFI::MemoryPointer.new(:uint8) })
+          ::SDL.GetRGB(num, self, *color)
         end
         color.map(&:read_uint8)
       end